Added rendering of pictures with additional information
